# Improve performance and reduce allocations for PropertyInjectionStrategy