… src/Ninject.Test/Unit/PropertyInjectionStrategyTests.cs