* Improves performance and reduce allocations for **PropertyInjectionStrategy**.
* Use **(I)ExceptionFormatter** as component in **PropertyInjectionStrategy**.

**Before:**

|                                                                      Method |         Mean |        Error |       StdDev | Gen 0/1k Op | Gen 1/1k Op | Gen 2/1k Op | Allocated Memory/Op |
----------------------------------------------------------------------------- |-------------:|-------------:|-------------:|------------:|------------:|------------:|--------------------:|
|    Activate_WithoutPropertyInjectionDirectivesAndWithPropertyValuesFullMatch | 276,960.0 ns | 1,659.527 ns | 1,471.128 ns |      3.9063 |      1.9531 |           - |             17481 B |
| Activate_WithoutPropertyInjectionDirectivesAndWithPropertyValuesPartialMatch | 187,869.7 ns | 1,512.608 ns | 1,340.887 ns |      2.6855 |      1.2207 |           - |             11880 B |
|          Activate_WithoutPropertyInjectionDirectivesAndWithoutPropertyValues |     358.7 ns |     3.609 ns |     3.376 ns |      0.0815 |           - |           - |               344 B |
|       Activate_WithPropertyInjectionDirectivesAndWithPropertyValuesFullMatch | 277,636.6 ns | 1,941.910 ns | 1,816.464 ns |      3.9063 |      1.9531 |           - |             17865 B |
|    Activate_WithPropertyInjectionDirectivesAndWithPropertyValuesPartialMatch | 194,956.4 ns | 1,258.220 ns | 1,176.940 ns |      3.1738 |      1.4648 |           - |             14121 B |
|             Activate_WithPropertyInjectionDirectivesAndWithoutPropertyValues |   6,240.7 ns |    49.278 ns |    46.095 ns |      1.2970 |           - |           - |              5456 B |             

**After:**

|                                                                      Method |         Mean |        Error |       StdDev | Gen 0/1k Op | Gen 1/1k Op | Gen 2/1k Op | Allocated Memory/Op |
----------------------------------------------------------------------------- |-------------:|-------------:|-------------:|------------:|------------:|------------:|--------------------:|
|    Activate_WithoutPropertyInjectionDirectivesAndWithPropertyValuesFullMatch | 263,434.6 ns | 3,778.842 ns | 3,534.731 ns |      3.9063 |      1.9531 |           - |             16641 B |
 Activate_WithoutPropertyInjectionDirectivesAndWithPropertyValuesPartialMatch | 172,033.5 ns | 1,559.241 ns | 1,382.226 ns |      2.6855 |      1.2207 |           - |             11288 B |
|          Activate_WithoutPropertyInjectionDirectivesAndWithoutPropertyValues |     127.9 ns |     1.475 ns |     1.231 ns |      0.0477 |           - |           - |               200 B |
|       Activate_WithPropertyInjectionDirectivesAndWithPropertyValuesFullMatch |     746.4 ns |     7.429 ns |     6.585 ns |      0.1268 |           - |           - |               536 B |
|    Activate_WithPropertyInjectionDirectivesAndWithPropertyValuesPartialMatch |   2,638.4 ns |    27.124 ns |    24.044 ns |      0.5379 |           - |           - |              2264 B |
|             Activate_WithPropertyInjectionDirectivesAndWithoutPropertyValues |   4,963.3 ns |    53.652 ns |    50.186 ns |      1.0757 |           - |           - |              4544 B |